Signup Flow: Change `null` errors to `undefined`
Possibly related to an issue where we can't move through the new site
flow in Firefox from the domain selection, there are a couple calls
to `SignupActions.submitSignupStep` that had `null` values for the
second `error` parameter and this diverges from every other call which
either passes `undefined`, `[]` or doesn't pass an `error` value.

I'm not sure if this fixes or even addresses the broken flow but
I also know that it's _good_ to harmonize all of the callsites.
Therefore I'd like to push out this change and see if it fixes
the problem in Firefox before doing much more investigation.